refactor(adapters): better env var handling and remove `is_complete`
Adapters can now have environment variables in the URL, the parameters
and the headers parts of their interface. This makes it much easier to
add support for additional services. Also, environment variables can now
also be functions as well as paths to elements in an adapter's schema.

Finally, the `is_complete` method has been removed from the adapter
interface in favour of using User autocmds. It's more flexible and far
more simpler to have the client fire an event to announce the request
has completed rather than checking every single response back from the
LLM for a particular pattern or string. Also, for some LLMs like Gemini,
it's impossible to determine when a request has completed using pattern
matching.