refactor(chat): messages can now store metadata
There's been a lot of awesome requests in recent weeks that have shown
the shortfall in the current approach to how the plugin parses messages
in the chat buffer.

Prior to this commit, when the user submitted a chat to the LLM, every
message in the buffer was parsed added to a table before being sent.
Now, we only take the most recent message and add it to a table in the
chat class. This allows us to embed options within the messages table
such as visibility and tags.

This approach does mean that you cannot edit previous messages in the
buffer.